requirements.txt
tensorboard
munch
opencv-python
h5py
scikit-learn
numpy
matplotlib
PyYAML